pyproject.toml
[project]
name = "student-os-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "pymupdf4llm>=0.0.17",
    "openai>=1.58.1",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.20",
]
